pyproject.toml
[project]
name = "streaming-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ayazkhanx", email = "ak96979590@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.8.4",
    "dotenv>=0.9.9",
    "openai-agents>=0.4.2",
]

[project.scripts]
streaming-agent = "streaming_agent:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
